+++ Client/N3Base/N3ShapeExtra.h
@@ -36,7 +36,7 @@ protected:
 	std::vector<__Rotation>	m_Rotations;
 
 public:
-	void RotateTo(int iPart, const __Vector3& vAxis, float fRadianToReach, float fRadianPerSec, bool bImmediately = false); // 원하는 파트를 축에 따라 지정한 각도까지 지정한 속도로 회전시킨다..
+	void RotateTo(size_t iPart, const __Vector3& vAxis, float fRadianToReach, float fRadianPerSec, bool bImmediately = false); // 원하는 파트를 축에 따라 지정한 각도까지 지정한 속도로 회전시킨다..
 
 	bool Load(HANDLE hFile);
 	void Tick(float fFrm);
